Improves styling of source editor and crashes
Prevents crashes from malformed json
