+++ apps/ui/src/views/campaign/editor/EmailEditor.css
@@ -93,9 +93,7 @@
     padding: 20px;
 }
 
-.source-editor {
-    --color-editor: #0d121e;
-    
+.source-editor {    
     background: var(--color-editor);
     border-radius: var(--border-radius);
     border: 1px solid #1a1f2b;
@@ -103,14 +101,6 @@
     padding: 0;
     display: flex;
     flex-direction: column;
-}
-
-.source-editor .monaco-editor {
-    --vscode-editor-background: var(--color-editor);
-    --vscode-editorStickyScroll-background: var(--color-editor);
-    --vscode-breadcrumb-background: var(--color-editor);
-    --vscode-editorGutter-background: var(--color-editor);
-    --vscode-editorMarkerNavigation-background: var(--color-editor);
 }
 
 .source-editor .ui-tabs {
